\name{Gabon Tree Data-package}
\alias{Gabon Tree Data-package}
\alias{Gabon Tree Data}
\docType{package}
\title{
\packageTitle{Gabon Tree Data}
}
\description{
\packageDescription{Gabon Tree Data}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Gabon Tree Data}
\packageIndices{Gabon Tree Data}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Gabon Tree Data}

Maintainer: \packageMaintainer{Gabon Tree Data}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
