\name{gabontreedata-package}
\alias{gabontreedata-package}
\alias{gabontreedata}
\docType{package}
\title{
\packageTitle{gabontreedata}
}
\description{
\packageDescription{gabontreedata}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gabontreedata}
\packageIndices{gabontreedata}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{gabontreedata}

Maintainer: \packageMaintainer{gabontreedata}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
